Connected Cliennt Model to DB With the drastic change in database architecture, the link between the database, model, and raw SQL has been established. The controller still needs to be connected.
While Node.js allows for the result of a require to point to another require, it appears that Express is not able to handle it out of the box.  Links had to be made to the DB module from the client model and not from the model index.

The raw SQL in edit-one-client had to be modified to allow for raw text entry of column names.

Most work was done in `server/models/client-model.js` where the client is able to updated a client record and return the updated client object to the caller.

The controller still needs to be completed.
 Changes to be committed:
	modified:   db/index.js
	modified:   db/psql/edit-one-client.sql
	modified:   db/psql/index.js
	modified:   server/controllers/index.js
	modified:   server/models/client-model.js
	modified:   server/models/index.js
refs #56